Start splitting up transformers for maintainability.
